Move corpus to rc-zip
