+++ rc-zip-sync/Cargo.toml
@@ -49,3 +49,4 @@ humansize = "2.1.3"
 indicatif = "0.17.7"
 test-log = { version = "0.2.14", default-features = false, features = ["tracing-subscriber", "trace"] }
 tracing-subscriber = { version = "0.3.18", features = ["env-filter"] }
+rc-zip = { version = "2.0.1", path = "../rc-zip", features = ["corpus"] }
