+++ rc-zip/Cargo.toml
@@ -32,4 +32,5 @@ miniz_oxide = { version = "0.7.1", optional = true }
 
 [features]
 deflate = ["dep:miniz_oxide"]
+corpus = []
